+++ packages/core/src/EntityManager.ts
@@ -617,7 +617,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
     em.validator.validateParams(data, 'insert data');
 
     if (em.eventManager.hasListeners(EventType.beforeUpsert, meta)) {
-      await em.eventManager.dispatchEvent(EventType.beforeUpsert, { entity: data, em }, meta);
+      await em.eventManager.dispatchEvent(EventType.beforeUpsert, { entity: data, em, meta }, meta);
     }
 
     const ret = await em.driver.nativeUpdate(entityName, where, data, {
@@ -640,7 +640,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
 
     if (!helper(entity).hasPrimaryKey()) {
       const pk = await this.driver.findOne(meta.className, where, {
-        fields: meta.hydrateProps.filter(p => !p.lazy && !(p.name in data!)).map(p => p.name) as EntityField<Entity>[],
+        fields: meta.comparableProps.filter(p => !p.lazy && !(p.name in data!)).map(p => p.name) as EntityField<Entity>[],
         ctx: em.transactionContext,
         convertCustomTypes: true,
       });
@@ -659,7 +659,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
     em.unitOfWork.registerManaged(entity, snapshot, { refresh: true });
 
     if (em.eventManager.hasListeners(EventType.afterUpsert, meta)) {
-      await em.eventManager.dispatchEvent(EventType.afterUpsert, { entity, em }, meta);
+      await em.eventManager.dispatchEvent(EventType.afterUpsert, { entity, em, meta }, meta);
     }
 
     return entity;
@@ -778,7 +778,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
     if (em.eventManager.hasListeners(EventType.beforeUpsert, meta)) {
       for (const dto of data) {
         const entity = entitiesByData.get(dto) ?? dto;
-        await em.eventManager.dispatchEvent(EventType.beforeUpsert, { entity, em }, meta);
+        await em.eventManager.dispatchEvent(EventType.beforeUpsert, { entity, em, meta }, meta);
       }
     }
 
@@ -821,7 +821,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
       }
 
       const pks = await this.driver.find(meta.className, { $or: [...loadPK.values()] as Dictionary[] }, {
-        fields: meta.hydrateProps.filter(p => !p.lazy && !(p.name in data![0] && !add.has(p.name))).map(p => p.name),
+        fields: meta.comparableProps.filter(p => !p.lazy && !(p.name in data![0] && !add.has(p.name))).map(p => p.name),
         ctx: em.transactionContext,
         convertCustomTypes: true,
       });
@@ -854,7 +854,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
 
     if (em.eventManager.hasListeners(EventType.afterUpsert, meta)) {
       for (const [entity] of entities) {
-        await em.eventManager.dispatchEvent(EventType.afterUpsert, { entity, em }, meta);
+        await em.eventManager.dispatchEvent(EventType.afterUpsert, { entity, em, meta }, meta);
       }
     }