+++ tests/issues/GH3271.test.ts
@@ -8,7 +8,10 @@ let orm: MikroORM;
 
 beforeAll(async () => orm = await initORMPostgreSql());
 beforeEach(async () => orm.schema.clearDatabase());
-afterAll(() => orm.close(true));
+afterAll(async () => {
+  await orm.schema.dropDatabase();
+  await orm.close(true);
+});
 
 test('test nested find with repository', async () => {
   const mock = mockLogger(orm, ['query', 'query-params']);